requirements.txt
Django==1.7
argparse==1.2.1
django-filter==0.7
django-grappelli==2.5.3
django-guardian==1.2.4
djangorestframework==2.4.2
six==1.7.3
wsgiref==0.1.2
